fixed hardcoded r5r jar path in setup_r5()
Although this probably wouldn't cause much problems to the majority of
the users, hardcoding the path to point to .libPaths()[1] could lead to
issues if r5r is not in this library.
I, for instance, wanted to test the package without installing it, only
loading via devtools, so the correct r5r's jar wasn't in that library,
but inside the dev's inst/jar directory.
Using system.file() fixes this issue.